Add more javadocs, package-privateness, and more efficient reading
