+++ core/http-auth-aws/src/main/java/software/amazon/awssdk/http/auth/aws/chunkedencoding/ChunkExtension.java
@@ -21,7 +21,13 @@ import software.amazon.awssdk.utils.Pair;
 /**
  * A functional interface for defining an extension of a chunk, where the extension is a key-value pair.
  * <p>
- * The extension usually depends on the chunk-data itself (checksum, signature, etc.), but is not required to.
+ * An extension usually depends on the chunk-data itself (checksum, signature, etc.), but is not required to.
+ * Per <a href="https://datatracker.ietf.org/doc/html/rfc7230#section-4.1.1">RFC-7230</a> The chunk-extension is defined as:
+ * <pre>
+ *     chunk-ext      = *( ";" chunk-ext-name [ "=" chunk-ext-val ] )
+ *     chunk-ext-name = token
+ *     chunk-ext-val  = token / quoted-string
+ * </pre>
  */
 @FunctionalInterface
 @SdkProtectedApi
